to flake8

+++ src/products.py
@@ -2,8 +2,8 @@ from typing import Any, Dict, List, Optional, Type
 
 from src.base_item import BaseItem
 from src.base_product import BaseProduct
-from src.print_mixin import PrintMixin
 from src.exceptions import ZeroQuantityError
+from src.print_mixin import PrintMixin
 
 
 class Product(BaseProduct, PrintMixin):
@@ -168,7 +168,7 @@ class Category(BaseItem):
         if not self.__products:
             return 0
         try:
-             total_price = sum(product.price for product in self.__products)
-             return total_price / len(self.__products)
+            total_price = sum(product.price for product in self.__products)
+            return float(total_price / len(self.__products))
         except ZeroDivisionError:
             return 0
